feat(sentry): 重构 Sentry 库并添加新功能
- 重构了 Sentry2 类,增加了新的成员函数
- 添加了 tosee_sentry 命名空间,实现了多个 Sentry 相关函数
- 修改了 SentryFactory 结构体和相关函数,优化了参数类型
- 更新了 sentry_i2c 和 sentry_uart 文件中的函数定义
- 调整了 SentryStream 相关的定义和实现
- 移除了未使用的 DOPRINTF 宏

+++ SentryFactory.h
@@ -19,7 +19,7 @@
 #include "sentry_uart.h"
 
 typedef struct {
-    uint32_t address;
+    uint8_t address;
     uint8_t device_id;
     int vision_max_type;
     int vision_qrcode_type;
@@ -31,7 +31,7 @@ typedef struct {
     sentry_qrcode_state_t *qrcode_state;
 } SentryFactory;
 
-void SentryFactory_Init(SentryFactory *factory, uint32_t address, uint8_t device_id,
+void SentryFactory_Init(SentryFactory *factory, uint8_t address, uint8_t device_id,
                         sentry_vision_state_t **vision_state, int vision_max_type,
                         int vision_qrcode_type);
